pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "google-genai>=1.32.0",
    "gunicorn>=23.0.0",
    "openai>=1.102.0",
    "python-dotenv>=1.1.1",
    "sqlitecloud>=0.0.84",
    "werkzeug>=3.1.3",
]
